fix(persistence): address review blockers on PR #1488
Three changes from review:

1. Migration path: update sqlx::migrate! in app/projects/{project_notes,
   project_manifests} from db/migrations to core/migrations; delete the
   Windows-hostile db/migrations symlink.

2. Audit write helpers: consolidate insert_audit_entry,
   insert_project_auto_transition_conn, and insert_resolution_audit into
   persistence_core::repositories::audit_writes; delete the duplicate in
   plans/audit_helpers.rs and the inline copy in targets/q_targets_mgmt;
   lifecycle/audit and targets/q_targets_mgmt now delegate to core.

3. Extract persistence_calibration in this PR (atomic domain extraction):
   calibration_assignment, calibration_tolerances, equipment, q_calibration
   move from persistence_targets to a new persistence_calibration crate.
   Evidence: equipment is consumed by app/calibration + app/inbox (import
   graph in .scratch.md), making it calibration-domain not targets-domain.
   Facade re-exports updated accordingly.

Tracks-Bead: astro-plan-gl58
