+++ Cargo.toml
@@ -26,6 +26,7 @@ members = [
   "crates/metadata/video",
   "crates/metadata/xisf",
   "crates/patterns",
+  "crates/persistence/calibration",
   "crates/persistence/core",
   "crates/persistence/db",
   "crates/persistence/inbox",
