+++ crates/persistence/calibration/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "persistence_core"
+name = "persistence_calibration"
 version = "0.1.0"
 edition.workspace = true
 license.workspace = true
@@ -8,18 +8,20 @@ license.workspace = true
 path = "src/lib.rs"
 
 [dependencies]
-audit_types = { path = "../../audit-types" }
+domain_core = { path = "../../domain/core" }
+persistence_core = { path = "../core" }
+sessions = { path = "../../sessions" }
 serde.workspace = true
 serde_json.workspace = true
 sqlx.workspace = true
 thiserror.workspace = true
 time.workspace = true
-tokio.workspace = true
 uuid.workspace = true
 
 [dev-dependencies]
-sea-query.workspace = true
+persistence_core = { path = "../core", features = [] }
 tempfile.workspace = true
+tokio.workspace = true
 
 [lints]
 workspace = true
