+++ crates/persistence/calibration/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright (C) 2024-2026 Sjors Robroek
+// SPDX-License-Identifier: AGPL-3.0-only
+
+//! Calibration domain persistence: equipment, assignments, tolerances, calibration queries.
+#![allow(clippy::doc_markdown)]
+
+pub mod repositories;
