+++ crates/persistence/calibration/src/repositories/equipment.rs
