+++ crates/persistence/calibration/src/repositories/mod.rs
@@ -1,0 +1,7 @@
+// Copyright (C) 2024-2026 Sjors Robroek
+// SPDX-License-Identifier: AGPL-3.0-only
+
+pub mod calibration_assignment;
+pub mod calibration_tolerances;
+pub mod equipment;
+pub mod q_calibration;
