--- crates/persistence/db/migrations
@@ -1,1 +1,0 @@
-../core/migrations
