+++ crates/persistence/targets/src/repositories/mod.rs
@@ -1,12 +1,8 @@
 // Copyright (C) 2024-2026 Sjors Robroek
 // SPDX-License-Identifier: AGPL-3.0-only
 
-pub mod calibration_assignment;
-pub mod calibration_tolerances;
-pub mod equipment;
 pub mod framing;
 pub mod inventory;
-pub mod q_calibration;
 pub mod q_desktop;
 pub mod q_resolver;
 pub mod q_targets_ingest;
